Extract HTML rendering logic into APIReadMe
