+++ Sources/GitHubTrendingRSSKit/Models/API/APIReadMe.swift
@@ -3,7 +3,10 @@
 //
 //
 
+import DocslothAsciidoctor
+import DocslothMarkdownItGFMCJKFriendly
 import Foundation
+import SwiftSoup
 
 public struct APIReadMe: Codable, Sendable {
     public enum FileType: CaseIterable {
@@ -101,4 +104,94 @@ public struct APIReadMe: Codable, Sendable {
     }
 
     public init() {}
+
+    public func makeHTML(supportedEmojis: [GitHubEmoji]) async throws -> String? {
+        guard let readMeHTML = try await renderHTML() else {
+            return nil
+        }
+        let normalizedReadMeHTML = normalizeEmojisInHTML(readMeHTML, supportedEmojis: supportedEmojis)
+        guard let parsedHTML = try? SwiftSoup.parse(normalizedReadMeHTML) else {
+            return normalizedReadMeHTML
+        }
+
+        guard let blobToRawRegex = try? NSRegularExpression(
+            pattern: "(https://github.com/[^/]+/[^/]+/)blob(/.+)",
+            options: []
+        ) else {
+            assertionFailure()
+            return normalizedReadMeHTML
+        }
+
+        let tagAttributesPairs = [
+            "a": ["href"],
+            "area": ["href"],
+            "img": ["src", "longdesc", "usemap"],
+            "link": ["href"],
+            "blockquote": ["cite"],
+        ]
+
+        for (tag, attributes) in tagAttributesPairs {
+            guard let elements = try? parsedHTML.getElementsByTag(tag) else {
+                continue
+            }
+            for element in elements {
+                for attribute in attributes {
+                    guard let url = try? element.attr(attribute).prefixDeleted(prefix: "/"),
+                          let baseURL = fileRootURL,
+                          var absoluteURL = URL(string: url, relativeTo: baseURL)?.absoluteString else {
+                        continue
+                    }
+                    if absoluteURL.hasSuffix(".svg"), tag == "img", attribute == "src" {
+                        absoluteURL += "?sanitize=true"
+                    }
+
+                    absoluteURL = blobToRawRegex.stringByReplacingMatches(
+                        in: absoluteURL,
+                        range: NSRange(absoluteURL.startIndex..., in: absoluteURL),
+                        withTemplate: "$1raw$2"
+                    )
+
+                    if absoluteURL != url {
+                        _ = try? element.attr(attribute, absoluteURL)
+                    }
+                }
+            }
+        }
+        return try? parsedHTML.body()?.html()
+    }
+
+    private func normalizeEmojisInHTML(_ readMeHTML: String, supportedEmojis: [GitHubEmoji]) -> String {
+        var normalized = readMeHTML
+        for emoji in supportedEmojis {
+            let target = ":" + emoji.id + ":"
+            normalized = normalized.replacingOccurrences(of: target, with: emoji.html)
+        }
+        return normalized
+    }
+
+    private func renderHTML() async throws -> String? {
+        guard let content else {
+            return nil
+        }
+        let html: String?
+        switch fileType {
+        case .markdown, .unknown:
+            html = try await DocslothManager.shared.markdownIt.convertToHTML(content)
+        case .asciiDoc:
+            html = try await DocslothManager.shared.asciidoctor.convertToHTML(content)
+        }
+        return html
+    }
 }
+
+private extension GitHubEmoji {
+    var html: String {
+        switch value {
+        case let .text(text):
+            return "<span>\(text)</span>"
+        case let .image(url):
+            return "<img alt='\(id)' src='\(url.absoluteString)' />"
+        }
+    }
+}
+
